fix: Form List warning (#24867)
* fix: Form List warning

* clean reaf
