+++ tests/utils.ts
@@ -8,4 +8,6 @@ export function resetMockDate() {
   MockDate.reset();
 }
 
-export const sleep = (timeout = 0) => new Promise(resolve => setTimeout(resolve, timeout));
+const globalTimeout = global.setTimeout;
+
+export const sleep = (timeout = 0) => new Promise(resolve => globalTimeout(resolve, timeout));
